refactor(media-library): extract background-task progress derivations
Moves the ~150-line cluster of pure progress derivations (proxy/transcription/
analysis/preparation counts, average progress, single-item stage label, and the
expandable per-item rows) into a self-contained useMediaTaskProgress hook that
reads its own store slices. Verbatim derivations.

media-library.tsx: 1694 -> 1546 lines (cumulative 1882 -> 1546, -18%).
